+++ src/client.ts
@@ -28,6 +28,7 @@ let extensions = {
 
 const STANZA_DISCONNECTED = 'stanzaDisconnected';
 const NO_LONGER_SUBSCRIBED = 'notify:no_longer_subscribed';
+const DUPLICATE_ID = 'notify:duplicate_id';
 const MAX_CHANNEL_REUSES = 10;
 
 export class Client extends EventEmitter {
@@ -48,6 +49,7 @@ export class Client extends EventEmitter {
 
   private boundStanzaDisconnect?: () => Promise<any>;
   private boundStanzaNoLongerSubscribed?: () => void;
+  private boundStanzaDuplicateId?: () => void;
 
   http: HttpClient;
   notifications!: NotificationsAPI;
@@ -159,9 +161,11 @@ export class Client extends EventEmitter {
 
     this.boundStanzaDisconnect = this.handleStanzaDisconnectedEvent.bind(this, stanza);
     this.boundStanzaNoLongerSubscribed = this.handleNoLongerSubscribed.bind(this, stanza);
+    this.boundStanzaDuplicateId = this.handleDuplicateId.bind(this, stanza);
 
     this.on(STANZA_DISCONNECTED, this.boundStanzaDisconnect);
     this.on(NO_LONGER_SUBSCRIBED, this.boundStanzaNoLongerSubscribed);
+    this.on(DUPLICATE_ID, this.boundStanzaDuplicateId);
 
     this.extensions.forEach(extension => {
       if (typeof extension.handleIq === 'function') {
@@ -182,6 +186,11 @@ export class Client extends EventEmitter {
     if (this.boundStanzaNoLongerSubscribed) {
       this.off(NO_LONGER_SUBSCRIBED, this.boundStanzaNoLongerSubscribed);
       this.boundStanzaNoLongerSubscribed = undefined;
+    }
+
+    if (this.boundStanzaDuplicateId) {
+      this.off(DUPLICATE_ID, this.boundStanzaDuplicateId);
+      this.boundStanzaDuplicateId = undefined;
     }
   }
 
@@ -241,6 +250,13 @@ export class Client extends EventEmitter {
     if (!this.reconnectOnNoLongerSubscribed) {
       this.autoReconnect = false;
     }
+  }
+
+  private handleDuplicateId (stanzaInstance: NamedAgent) {
+    this.logger.warn('duplicate_id event received, forcing hard reconnect', { stanzaInstanceId: stanzaInstance.id, channelId: stanzaInstance.channelId });
+    stanzaInstance.pinger?.stop();
+
+    this.hardReconnectRequired = true;
   }
 
   async disconnect () {